// src/app/models/product.model.ts
export interface Product {
  id: string;
  title: string;
  price: number;
  description: string;
  category: Category;
  images: string[];
  taxes?: number;
}

export interface Category {
  id: string;
  name: string;
  image: string;
}

export interface CreateProductDTO extends Omit<Product, 'id' | 'category'> {
  categoryId: number;
}

export interface UpdateProductDTO extends Partial<CreateProductDTO> { } //Partial aplica ? a todas las propiedades